Widgetテストのフォルダ構成を変更
