+++ test/m/2018-2019/regular/10/m_20181004_test.dart
@@ -2,210 +2,247 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mahjong_score/data/fu/fus.dart';
 import 'package:mahjong_score/data/yaku/yaku.dart';
 
-import '../../test_widget_util.dart';
-
-// 東3局 1本場：流局
-// expect(find.text('断么九'), findsOneWidget);
+import '../../../test_widget_util.dart';
 
 void main() {
-  group('2018/10/15 1試合目', () {
+  group('2018/10/04 1試合目', () {
     // 東1局 0本場：流局
-    testWidgets('東1局 0本場', (WidgetTester tester) async {
+    testWidgets('東1局 1本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
             YakuId.reach,
+            YakuId.tsumo,
           ],
-          dora: 1,
+          dora: 2,
           honba: 1,
           menzen: true,
-          tsumo: false,
-          oya: false,
+          tsumo: true,
+          oya: true,
           mentsu1: FuMentsu.shuntsu,
           mentsu2: FuMentsu.shuntsu,
           mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.chuchanpaiAnko,
-          atama: FuAtama.otakaze,
-          machi: FuMachi.ryanmen,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.penchan,
         ),
       );
 
       expect(find.text('立直'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
 
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('2900点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東2局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.honitsu,
-            YakuId.chitoitsu,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.none,
-          mentsu2: FuMentsu.none,
-          mentsu3: FuMentsu.none,
-          mentsu4: FuMentsu.none,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('混一色'), findsOneWidget);
-      expect(find.text('七対子'), findsOneWidget);
-
-      expect(find.text('5翻'),
+      expect(find.text('4翻'),
           findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('25符'), findsOneWidget);
+      expect(find.text('30符'), findsOneWidget);
 
-      expect(find.text('12000点'), findsOneWidget);
+      expect(find.text('4100点 all'), findsOneWidget);
 
-      expect(find.text('基本：25'), findsOneWidget);
+      expect(find.text('基本：20'), findsOneWidget);
       expect(find.text('面子：0'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
+      expect(find.text('待ち：2'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
     });
 
-    testWidgets('東2局 1本場', (WidgetTester tester) async {
+    testWidgets('東1局 2本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
             YakuId.reach,
             YakuId.ippatsu,
             YakuId.tsumo,
-            YakuId.tonBa,
-            YakuId.tonKaze,
           ],
           dora: 1,
-          honba: 1,
+          honba: 2,
           menzen: true,
           tsumo: true,
-          oya: true,
+          oya: false,
           mentsu1: FuMentsu.shuntsu,
           mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.yaochuhaiAnko,
+          mentsu3: FuMentsu.shuntsu,
           mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.otakaze,
-          machi: FuMachi.ryanmen,
+          atama: FuAtama.suhai,
+          machi: FuMachi.kanchan,
         ),
       );
 
       expect(find.text('立直'), findsOneWidget);
       expect(find.text('一発'), findsOneWidget);
       expect(find.text('門前自摸'), findsOneWidget);
-      expect(find.text('東（場）'), findsOneWidget);
-      expect(find.text('東（風）'), findsOneWidget);
 
-      expect(find.text('6翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('4翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('30符'), findsOneWidget);
 
-      expect(find.text('6100点 all'), findsOneWidget);
+      expect(find.text('子：2200点'), findsOneWidget);
+      expect(find.text('親：4200点'), findsOneWidget);
 
       expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：8'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('待ち：2'), findsOneWidget);
       expect(find.text('ツモ：2'), findsOneWidget);
     });
 
-    testWidgets('東2局 2本場', (WidgetTester tester) async {
+    testWidgets('東2局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.pei,
+            YakuId.sha,
+            YakuId.honitsu,
           ],
-          dora: 5,
-          honba: 2,
+          dora: 1,
+          honba: 0,
           menzen: false,
-          tsumo: false,
+          tsumo: true,
           oya: false,
           mentsu1: FuMentsu.shuntsu,
           mentsu2: FuMentsu.shuntsu,
           mentsu3: FuMentsu.yaochuhaiMinko,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
+          mentsu4: FuMentsu.yaochuhaiMinko,
+          atama: FuAtama.sangenpai,
           machi: FuMachi.tanki,
         ),
       );
 
-      expect(find.text('北'), findsOneWidget);
+      expect(find.text('西'), findsOneWidget);
+      expect(find.text('混一色'), findsOneWidget);
 
-      expect(find.text('6翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
+      expect(find.text('4翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('40符'), findsOneWidget);
 
-      expect(find.text('12600点'), findsOneWidget);
+      expect(find.text('子：2000点'), findsOneWidget);
+      expect(find.text('親：4000点'), findsOneWidget);
 
       expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('面子：8'), findsOneWidget);
+      expect(find.text('アタマ：2'), findsOneWidget);
       expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
     });
 
-    // 東3局 0本場：流局
-
-    testWidgets('東3局 1本場', (WidgetTester tester) async {
+    testWidgets('東3局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.hatsu,
+            YakuId.reach,
+            YakuId.tsumo,
           ],
           dora: 0,
-          honba: 1,
-          menzen: false,
+          honba: 0,
+          menzen: true,
+          tsumo: true,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.yaochuhaiAnko,
+          atama: FuAtama.sangenpai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+
+      expect(find.text('2翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('40符'), findsOneWidget);
+
+      expect(find.text('子：700点'), findsOneWidget);
+      expect(find.text('親：1300点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：8'), findsOneWidget);
+      expect(find.text('アタマ：2'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
+    });
+
+    testWidgets('東4局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.tsumo,
+          ],
+          dora: 2,
+          honba: 0,
+          menzen: true,
           tsumo: true,
           oya: true,
           mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.chuchanpaiMinko,
-          mentsu3: FuMentsu.yaochuhaiMinko,
-          mentsu4: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.chuchanpaiAnko,
           atama: FuAtama.suhai,
           machi: FuMachi.ryanmen,
         ),
       );
 
-      expect(find.text('發'), findsOneWidget);
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
 
-      expect(find.text('1翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('4翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('30符'), findsOneWidget);
 
-      expect(find.text('600点 all'), findsOneWidget);
+      expect(find.text('4000点 all'), findsOneWidget);
 
       expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：6'), findsOneWidget);
+      expect(find.text('面子：4'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
       expect(find.text('ツモ：2'), findsOneWidget);
     });
 
-    testWidgets('東3局 2本場', (WidgetTester tester) async {
+    testWidgets('東4局 1本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.chun,
+          ],
+          dora: 2,
+          honba: 1,
+          menzen: false,
+          tsumo: false,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.yaochuhaiMinko,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('中'), findsOneWidget);
+
+      expect(find.text('3翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('6100点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：4'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    // 東4局 2本場：流局
+
+    testWidgets('南1局 2本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
             YakuId.reach,
             YakuId.pinfu,
           ],
-          dora: 4,
-          honba: 2,
+          dora: 1,
+          honba: 3,
           menzen: true,
           tsumo: false,
           oya: false,
@@ -221,11 +258,11 @@ void main() {
       expect(find.text('立直'), findsOneWidget);
       expect(find.text('平和'), findsOneWidget);
 
-      expect(find.text('6翻'),
+      expect(find.text('3翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('30符'), findsOneWidget);
 
-      expect(find.text('12600点'), findsOneWidget);
+      expect(find.text('4800点'), findsOneWidget);
 
       expect(find.text('基本：30'), findsOneWidget);
       expect(find.text('面子：0'), findsOneWidget);
@@ -234,11 +271,203 @@ void main() {
       expect(find.text('ツモ：0'), findsOneWidget);
     });
 
-    testWidgets('東4局 0本場', (WidgetTester tester) async {
+    // 南2局 0本場：流局
+
+    testWidgets('南2局 1本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
+            YakuId.chun,
+          ],
+          dora: 2,
+          honba: 1,
+          menzen: false,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.yaochuhaiMinko,
+          atama: FuAtama.otakaze,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('中'), findsOneWidget);
+
+      expect(find.text('3翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('4200点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：4'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('南3局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.pinfu,
+            YakuId.tsumo,
+          ],
+          dora: 2,
+          honba: 0,
+          menzen: true,
+          tsumo: true,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('平和'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+
+      expect(find.text('5翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('20符'), findsOneWidget);
+
+      expect(find.text('4000点 all'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('南3局 1本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.pinfu,
+            YakuId.tsumo,
+            YakuId.ippatsu,
+          ],
+          dora: 1,
+          honba: 1,
+          menzen: true,
+          tsumo: true,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('平和'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+      expect(find.text('一発'), findsOneWidget);
+
+      expect(find.text('5翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('20符'), findsOneWidget);
+
+      expect(find.text('子：2100点'), findsOneWidget);
+      expect(find.text('親：4100点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('南4局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+          ],
+          dora: 2,
+          honba: 0,
+          menzen: true,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.yaochuhaiAnko,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+
+      expect(find.text('3翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('40符'), findsOneWidget);
+
+      expect(find.text('5200点'), findsOneWidget);
+
+      expect(find.text('基本：30'), findsOneWidget);
+      expect(find.text('面子：8'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+  });
+
+  group('2018/10/04 2試合目', () {
+    testWidgets('東1局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
             YakuId.tanyao,
+          ],
+          dora: 2,
+          honba: 0,
+          menzen: true,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.kanchan,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('断么九'), findsOneWidget);
+
+      expect(find.text('4翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('40符'), findsOneWidget);
+
+      expect(find.text('8000点'), findsOneWidget);
+
+      expect(find.text('基本：30'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：2'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('東2局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
             YakuId.pinfu,
             YakuId.tsumo,
           ],
@@ -256,7 +485,7 @@ void main() {
         ),
       );
 
-      expect(find.text('断么九'), findsOneWidget);
+      expect(find.text('立直'), findsOneWidget);
       expect(find.text('平和'), findsOneWidget);
       expect(find.text('門前自摸'), findsOneWidget);
 
@@ -274,56 +503,213 @@ void main() {
       expect(find.text('ツモ：0'), findsOneWidget);
     });
 
-    testWidgets('南1局 0本場', (WidgetTester tester) async {
+    testWidgets('東3局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
             YakuId.reach,
-            YakuId.ipeiko,
+            YakuId.pinfu,
             YakuId.tsumo,
           ],
-          dora: 5,
+          dora: 2,
           honba: 0,
           menzen: true,
           tsumo: true,
-          oya: false,
+          oya: true,
           mentsu1: FuMentsu.shuntsu,
           mentsu2: FuMentsu.shuntsu,
           mentsu3: FuMentsu.shuntsu,
           mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.baKaze,
+          atama: FuAtama.suhai,
           machi: FuMachi.ryanmen,
         ),
       );
 
       expect(find.text('立直'), findsOneWidget);
-      expect(find.text('一盃口'), findsOneWidget);
+      expect(find.text('平和'), findsOneWidget);
       expect(find.text('門前自摸'), findsOneWidget);
 
-      expect(find.text('8翻'),
+      expect(find.text('5翻'),
           findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
+      expect(find.text('20符'), findsOneWidget);
 
-      expect(find.text('子：4000点'), findsOneWidget);
-      expect(find.text('親：8000点'), findsOneWidget);
+      expect(find.text('4000点 all'), findsOneWidget);
 
       expect(find.text('基本：20'), findsOneWidget);
       expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：2'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('東3局 1本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.pinfu,
+            YakuId.tanyao,
+            YakuId.tsumo,
+            YakuId.ippatsu,
+          ],
+          dora: 2,
+          honba: 1,
+          menzen: true,
+          tsumo: true,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('平和'), findsOneWidget);
+      expect(find.text('断么九'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+      expect(find.text('一発'), findsOneWidget);
+
+      expect(find.text('7翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('20符'), findsOneWidget);
+
+      expect(find.text('6100点 all'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    // 東3局 2本場：流局
+    // 東4局 3本場：流局
+
+    testWidgets('南1局 4本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.tanyao,
+          ],
+          dora: 1,
+          honba: 4,
+          menzen: false,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('断么九'), findsOneWidget);
+
+      expect(find.text('2翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('3200点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
     });
 
     testWidgets('南2局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.reach,
             YakuId.tanyao,
-            YakuId.pinfu,
           ],
           dora: 2,
           honba: 0,
+          menzen: false,
+          tsumo: true,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('断么九'), findsOneWidget);
+
+      expect(find.text('3翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('2000点 all'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
+    });
+
+    testWidgets('南2局 1本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.tsumo,
+            YakuId.tanyao,
+          ],
+          dora: 1,
+          honba: 1,
+          menzen: true,
+          tsumo: true,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.chuchanpaiAnko,
+          atama: FuAtama.suhai,
+          machi: FuMachi.kanchan,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+      expect(find.text('断么九'), findsOneWidget);
+
+      expect(find.text('4翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('4100点 all'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：4'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：2'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
+    });
+
+    // 南2局 2本場：流局
+
+    testWidgets('南2局 3本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.ippatsu,
+            YakuId.pinfu,
+            YakuId.ipeiko,
+          ],
+          dora: 1,
+          honba: 3,
           menzen: true,
           tsumo: false,
           oya: true,
@@ -337,14 +723,15 @@ void main() {
       );
 
       expect(find.text('立直'), findsOneWidget);
-      expect(find.text('断么九'), findsOneWidget);
+      expect(find.text('一発'), findsOneWidget);
       expect(find.text('平和'), findsOneWidget);
+      expect(find.text('一盃口'), findsOneWidget);
 
       expect(find.text('5翻'),
           findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('30符'), findsOneWidget);
 
-      expect(find.text('12000点'), findsOneWidget);
+      expect(find.text('12900点'), findsOneWidget);
 
       expect(find.text('基本：30'), findsOneWidget);
       expect(find.text('面子：0'), findsOneWidget);
@@ -353,19 +740,21 @@ void main() {
       expect(find.text('ツモ：0'), findsOneWidget);
     });
 
-    testWidgets('南2局 1本場', (WidgetTester tester) async {
+    testWidgets('南2局 4本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.chun,
+            YakuId.reach,
+            YakuId.ippatsu,
+            YakuId.pinfu,
           ],
-          dora: 0,
-          honba: 1,
-          menzen: false,
+          dora: 3,
+          honba: 4,
+          menzen: true,
           tsumo: false,
           oya: false,
           mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.yaochuhaiMinko,
+          mentsu2: FuMentsu.shuntsu,
           mentsu3: FuMentsu.shuntsu,
           mentsu4: FuMentsu.shuntsu,
           atama: FuAtama.suhai,
@@ -373,16 +762,18 @@ void main() {
         ),
       );
 
-      expect(find.text('中'), findsOneWidget);
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('一発'), findsOneWidget);
+      expect(find.text('平和'), findsOneWidget);
 
-      expect(find.text('1翻'),
+      expect(find.text('6翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('30符'), findsOneWidget);
 
-      expect(find.text('1300点'), findsOneWidget);
+      expect(find.text('13200点'), findsOneWidget);
 
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
+      expect(find.text('基本：30'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
       expect(find.text('ツモ：0'), findsOneWidget);
@@ -392,234 +783,13 @@ void main() {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.hatsu,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.yaochuhaiMinko,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('發'), findsOneWidget);
-
-      expect(find.text('1翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('1500点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南3局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.hatsu,
-          ],
-          dora: 1,
-          honba: 1,
-          menzen: false,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.yaochuhaiMinko,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('發'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('子：600点'), findsOneWidget);
-      expect(find.text('親：1100点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    testWidgets('南4局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.pinfu,
-          ],
-          dora: 3,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('平和'), findsOneWidget);
-
-      expect(find.text('4翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('8000点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-  });
-
-  group('2018/10/15 2試合目', () {
-    testWidgets('東1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
             YakuId.reach,
-            YakuId.ippatsu,
           ],
-          dora: 2,
+          dora: 1,
           honba: 0,
           menzen: true,
           tsumo: false,
           oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.kanchan,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('一発'), findsOneWidget);
-
-      expect(find.text('4翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('8000点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    // 東2局 0本場：流局
-
-    testWidgets('東3局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.haku,
-            YakuId.toitoiho,
-            YakuId.sananko,
-          ],
-          dora: 3,
-          honba: 1,
-          menzen: false,
-          tsumo: true,
-          oya: true,
-          mentsu1: FuMentsu.chuchanpaiAnko,
-          mentsu2: FuMentsu.chuchanpaiAnko,
-          mentsu3: FuMentsu.chuchanpaiAnko,
-          mentsu4: FuMentsu.yaochuhaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
-        ),
-      );
-
-      expect(find.text('白'), findsOneWidget);
-      expect(find.text('対々和'), findsOneWidget);
-      expect(find.text('三暗刻'), findsOneWidget);
-
-      expect(find.text('8翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('8100点 all'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：16'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    testWidgets('東3局 2本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-          ],
-          dora: 1,
-          honba: 2,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('2600点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東4局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.sanshokudojun,
-          ],
-          dora: 2,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: true,
           mentsu1: FuMentsu.shuntsu,
           mentsu2: FuMentsu.shuntsu,
           mentsu3: FuMentsu.shuntsu,
@@ -629,241 +799,17 @@ void main() {
         ),
       );
 
-      expect(find.text('三色同順'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('5800点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：2'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東4局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.hatsu,
-            YakuId.chun,
-          ],
-          dora: 1,
-          honba: 1,
-          menzen: false,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.yaochuhaiAnko,
-          mentsu3: FuMentsu.chuchanpaiMinko,
-          mentsu4: FuMentsu.yaochuhaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.tanki,
-        ),
-      );
-
-      expect(find.text('發'), findsOneWidget);
-      expect(find.text('中'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('8000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：14'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    // 東4局 2本場：流局
-    // 東4局 3本場：流局
-
-    testWidgets('東4局 4本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-            YakuId.tsumo,
-          ],
-          dora: 1,
-          honba: 4,
-          menzen: true,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.kanchan,
-        ),
-      );
-
-      expect(find.text('門前自摸'), findsOneWidget);
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('子：1400点'), findsOneWidget);
-      expect(find.text('親：2400点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    testWidgets('南1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.chuchanpaiMinko,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
+      expect(find.text('立直'), findsOneWidget);
 
       expect(find.text('2翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
+      expect(find.text('40符'), findsOneWidget);
 
-      expect(find.text('2000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：2'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南3局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-          ],
-          dora: 3,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.chuchanpaiMinko,
-          mentsu3: FuMentsu.chuchanpaiAnko,
-          mentsu4: FuMentsu.chuchanpaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('4翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('12000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：8'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    // 南3局 1本場：流局
-
-    testWidgets('南3局 2本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.chitoitsu,
-          ],
-          dora: 1,
-          honba: 2,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.none,
-          mentsu2: FuMentsu.none,
-          mentsu3: FuMentsu.none,
-          mentsu4: FuMentsu.none,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('七対子'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('25符'), findsOneWidget);
-
-      expect(find.text('5400点'), findsOneWidget);
-
-      expect(find.text('基本：25'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南3局 3本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.pinfu,
-            YakuId.tanyao,
-          ],
-          dora: 0,
-          honba: 3,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('平和'), findsOneWidget);
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('4800点'), findsOneWidget);
+      expect(find.text('2600点'), findsOneWidget);
 
       expect(find.text('基本：30'), findsOneWidget);
       expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('アタマ：2'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
       expect(find.text('ツモ：0'), findsOneWidget);
     });
@@ -872,32 +818,32 @@ void main() {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.haku,
+            YakuId.sananko,
           ],
-          dora: 2,
+          dora: 0,
           honba: 0,
           menzen: true,
           tsumo: false,
           oya: false,
-          mentsu1: FuMentsu.yaochuhaiMinko,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.chuchanpaiAnko,
+          mentsu3: FuMentsu.yaochuhaiAnko,
+          mentsu4: FuMentsu.yaochuhaiAnko,
           atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
+          machi: FuMachi.ryanmen,
         ),
       );
 
-      expect(find.text('白'), findsOneWidget);
+      expect(find.text('三暗刻'), findsOneWidget);
 
-      expect(find.text('3翻'),
+      expect(find.text('2翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
+      expect(find.text('50符'), findsOneWidget);
 
-      expect(find.text('5200点'), findsOneWidget);
+      expect(find.text('3200点'), findsOneWidget);
 
       expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
+      expect(find.text('面子：20'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
       expect(find.text('ツモ：0'), findsOneWidget);
