+++ test/m/2018-2019/regular/10/m_20181008_test.dart
@@ -2,20 +2,319 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mahjong_score/data/fu/fus.dart';
 import 'package:mahjong_score/data/yaku/yaku.dart';
 
-import '../../test_widget_util.dart';
-
-// 東3局 1本場：流局
-// expect(find.text('断么九'), findsOneWidget);
+import '../../../test_widget_util.dart';
 
 void main() {
-  group('2018/10/19 1試合目', () {
+  group('2018/10/08 1試合目', () {
     testWidgets('東1局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.ipeiko,
+          ],
+          dora: 4,
+          honba: 0,
+          menzen: true,
+          tsumo: false,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.penchan,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('一盃口'), findsOneWidget);
+
+      expect(find.text('6翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('40符'), findsOneWidget);
+
+      expect(find.text('18000点'), findsOneWidget);
+
+      expect(find.text('基本：30'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：2'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    // 東1局 1本場：流局
+
+    testWidgets('東2局 2本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.tanyao,
+            YakuId.sanshokudojun,
+          ],
+          dora: 1,
+          honba: 2,
+          menzen: false,
+          tsumo: false,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('断么九'), findsOneWidget);
+      expect(find.text('三色同順'), findsOneWidget);
+
+      expect(find.text('3翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('6400点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('東2局 3本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.tsumo,
+          ],
+          dora: 3,
+          honba: 3,
+          menzen: true,
+          tsumo: true,
+          oya: true,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.yaochuhaiAnko,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+
+      expect(find.text('5翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('4300点 all'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：8'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
+    });
+
+    testWidgets('東2局 4本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.tonBa,
+          ],
+          dora: 1,
+          honba: 4,
+          menzen: true,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.chuchanpaiAnko,
+          mentsu4: FuMentsu.yaochuhaiAnko,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('東（場）'), findsOneWidget);
+
+      expect(find.text('2翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('50符'), findsOneWidget);
+
+      expect(find.text('4400点'), findsOneWidget);
+
+      expect(find.text('基本：30'), findsOneWidget);
+      expect(find.text('面子：12'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('東3局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.hatsu,
+          ],
+          dora: 0,
+          honba: 0,
+          menzen: false,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.yaochuhaiMinko,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('發'), findsOneWidget);
+
+      expect(find.text('1翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('1000点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：4'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    // 東4局 0本場：流局
+
+    testWidgets('東4局 1本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.honitsu,
+            YakuId.toitoiho,
+            YakuId.rinshankaiho,
+          ],
+          dora: 0,
+          honba: 1,
+          menzen: false,
+          tsumo: true,
+          oya: false,
+          mentsu1: FuMentsu.chuchanpaiAnko,
+          mentsu2: FuMentsu.chuchanpaiAnko,
+          mentsu3: FuMentsu.chuchanpaiMinko,
+          mentsu4: FuMentsu.yaochuhaiMinko,
+          atama: FuAtama.baKaze,
+          machi: FuMachi.shanpon,
+        ),
+      );
+
+      expect(find.text('混一色'), findsOneWidget);
+      expect(find.text('対々和'), findsOneWidget);
+      expect(find.text('嶺上開花'), findsOneWidget);
+
+      expect(find.text('5翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('40符'), findsOneWidget);
+
+      expect(find.text('子：2100点'), findsOneWidget);
+      expect(find.text('親：4100点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：14'), findsOneWidget);
+      expect(find.text('アタマ：2'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
+    });
+
+    // 南1局 0本場：流局
+
+    testWidgets('南2局 1本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.pinfu,
+            YakuId.tsumo,
+          ],
+          dora: 3,
+          honba: 1,
+          menzen: true,
+          tsumo: true,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('平和'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+
+      expect(find.text('6翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('20符'), findsOneWidget);
+
+      expect(find.text('子：3100点'), findsOneWidget);
+      expect(find.text('親：6100点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('南3局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.tanyao,
+          ],
+          dora: 0,
+          honba: 0,
+          menzen: false,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.suhai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('断么九'), findsOneWidget);
+
+      expect(find.text('1翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('1000点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+
+    testWidgets('南4局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
             YakuId.pinfu,
           ],
-          dora: 0,
+          dora: 2,
           honba: 0,
           menzen: true,
           tsumo: false,
@@ -31,15 +330,52 @@ void main() {
 
       expect(find.text('平和'), findsOneWidget);
 
-      expect(find.text('1翻'),
+      expect(find.text('3翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('30符'), findsOneWidget);
 
-      expect(find.text('1000点'), findsOneWidget);
+      expect(find.text('3900点'), findsOneWidget);
 
       expect(find.text('基本：30'), findsOneWidget);
       expect(find.text('面子：0'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
+    });
+  });
+
+  group('2018/10/08 2試合目', () {
+    testWidgets('東1局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+          ],
+          dora: 2,
+          honba: 0,
+          menzen: true,
+          tsumo: false,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.shuntsu,
+          atama: FuAtama.sangenpai,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+
+      expect(find.text('3翻'),
+          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('40符'), findsOneWidget);
+
+      expect(find.text('5200点'), findsOneWidget);
+
+      expect(find.text('基本：30'), findsOneWidget);
+      expect(find.text('面子：0'), findsOneWidget);
+      expect(find.text('アタマ：2'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
       expect(find.text('ツモ：0'), findsOneWidget);
     });
@@ -81,348 +417,55 @@ void main() {
       expect(find.text('ツモ：0'), findsOneWidget);
     });
 
-    // 東3局 0本場：流局
-
-    testWidgets('東4局 1本場', (WidgetTester tester) async {
+    testWidgets('東3局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.tanyao,
+            YakuId.reach,
+            YakuId.tsumo,
           ],
-          dora: 0,
-          honba: 1,
-          menzen: false,
+          dora: 4,
+          honba: 0,
+          menzen: true,
           tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.chuchanpaiAnko,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('1翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('子：400点'), findsOneWidget);
-      expect(find.text('親：600点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    testWidgets('南1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tonKaze,
-            YakuId.toitoiho,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.chuchanpaiMinko,
-          mentsu2: FuMentsu.chuchanpaiAnko,
-          mentsu3: FuMentsu.yaochuhaiMinko,
-          mentsu4: FuMentsu.yaochuhaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
-        ),
-      );
-
-      expect(find.text('東（風）'), findsOneWidget);
-      expect(find.text('対々和'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('7700点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：14'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南1局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-          ],
-          dora: 1,
-          honba: 1,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('2300点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南2局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.nanBa,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
           oya: false,
           mentsu1: FuMentsu.shuntsu,
           mentsu2: FuMentsu.shuntsu,
           mentsu3: FuMentsu.chuchanpaiAnko,
-          mentsu4: FuMentsu.yaochuhaiMinko,
+          mentsu4: FuMentsu.yaochuhaiAnko,
           atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
+          machi: FuMachi.shanpon,
         ),
       );
 
-      expect(find.text('南（場）'), findsOneWidget);
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
 
-      expect(find.text('1翻'),
+      expect(find.text('6翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('1000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：8'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南3局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.nanBa,
-            YakuId.nanKaze,
-            YakuId.haku,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.yaochuhaiAnko,
-          mentsu4: FuMentsu.yaochuhaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('南（場）'), findsOneWidget);
-      expect(find.text('南（風）'), findsOneWidget);
-      expect(find.text('白'), findsOneWidget);
-
-      expect(find.text('4翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('40符'), findsOneWidget);
 
-      expect(find.text('8000点'), findsOneWidget);
+      expect(find.text('子：3000点'), findsOneWidget);
+      expect(find.text('親：6000点'), findsOneWidget);
 
       expect(find.text('基本：20'), findsOneWidget);
       expect(find.text('面子：12'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南4局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.kanchan,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('1翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('1000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-  });
-
-  group('2018/10/19 2試合目', () {
-    testWidgets('東1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: false,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.chuchanpaiAnko,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.kanchan,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('1翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('子：300点'), findsOneWidget);
-      expect(find.text('親：500点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
       expect(find.text('ツモ：2'), findsOneWidget);
     });
 
-    testWidgets('東2局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.hatsu,
-            YakuId.toitoiho,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: false,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.chuchanpaiAnko,
-          mentsu2: FuMentsu.yaochuhaiAnko,
-          mentsu3: FuMentsu.yaochuhaiMinko,
-          mentsu4: FuMentsu.chuchanpaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
-        ),
-      );
+    // 東4局 0本場：流局
 
-      expect(find.text('發'), findsOneWidget);
-      expect(find.text('対々和'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('子：1300点'), findsOneWidget);
-      expect(find.text('親：2600点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：18'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    // 東3局 0本場：流局
-
-    testWidgets('東3局 1本場', (WidgetTester tester) async {
+    testWidgets('南1局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
             YakuId.tanyao,
-          ],
-          dora: 1,
-          honba: 1,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.chuchanpaiMinko,
-          mentsu4: FuMentsu.chuchanpaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.shanpon,
-        ),
-      );
-
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('2300点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東4局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-            YakuId.pinfu,
           ],
           dora: 3,
           honba: 0,
-          menzen: true,
+          menzen: false,
           tsumo: false,
           oya: false,
           mentsu1: FuMentsu.shuntsu,
@@ -435,57 +478,18 @@ void main() {
       );
 
       expect(find.text('断么九'), findsOneWidget);
-      expect(find.text('平和'), findsOneWidget);
 
-      expect(find.text('5翻'),
+      expect(find.text('4翻'),
           findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('30符'), findsOneWidget);
 
       expect(find.text('8000点'), findsOneWidget);
 
-      expect(find.text('基本：30'), findsOneWidget);
+      expect(find.text('基本：20'), findsOneWidget);
       expect(find.text('面子：0'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
       expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    // 南1局 0本場：流局
-
-    testWidgets('南1局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.nanBa,
-          ],
-          dora: 0,
-          honba: 1,
-          menzen: false,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.yaochuhaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('南（場）'), findsOneWidget);
-
-      expect(find.text('1翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('子：400点'), findsOneWidget);
-      expect(find.text('親：600点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
     });
 
     testWidgets('南2局 0本場', (WidgetTester tester) async {
@@ -523,22 +527,21 @@ void main() {
       expect(find.text('ツモ：0'), findsOneWidget);
     });
 
-    // 南3局 0本場：流局
-
-    testWidgets('南3局 1本場', (WidgetTester tester) async {
+    testWidgets('南3局 0本場', (WidgetTester tester) async {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
             YakuId.chun,
+            YakuId.nanBa,
           ],
           dora: 0,
-          honba: 1,
+          honba: 0,
           menzen: false,
           tsumo: false,
           oya: false,
-          mentsu1: FuMentsu.shuntsu,
+          mentsu1: FuMentsu.yaochuhaiMinko,
           mentsu2: FuMentsu.yaochuhaiAnko,
-          mentsu3: FuMentsu.yaochuhaiMinko,
+          mentsu3: FuMentsu.chuchanpaiMinko,
           mentsu4: FuMentsu.shuntsu,
           atama: FuAtama.suhai,
           machi: FuMachi.shanpon,
@@ -546,15 +549,16 @@ void main() {
       );
 
       expect(find.text('中'), findsOneWidget);
+      expect(find.text('南（場）'), findsOneWidget);
 
-      expect(find.text('1翻'),
+      expect(find.text('2翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
       expect(find.text('40符'), findsOneWidget);
 
-      expect(find.text('1600点'), findsOneWidget);
+      expect(find.text('2600点'), findsOneWidget);
 
       expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：12'), findsOneWidget);
+      expect(find.text('面子：14'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
       expect(find.text('ツモ：0'), findsOneWidget);
@@ -564,36 +568,35 @@ void main() {
       await tester.pumpWidget(
         createTestProviderScope(
           yakus: [
-            YakuId.nanBa,
+            YakuId.reach,
           ],
-          dora: 0,
+          dora: 1,
           honba: 0,
-          menzen: false,
-          tsumo: true,
+          menzen: true,
+          tsumo: false,
           oya: false,
           mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.yaochuhaiMinko,
-          mentsu3: FuMentsu.yaochuhaiMinko,
-          mentsu4: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.yaochuhaiAnko,
           atama: FuAtama.suhai,
           machi: FuMachi.ryanmen,
         ),
       );
 
-      expect(find.text('南（場）'), findsOneWidget);
+      expect(find.text('立直'), findsOneWidget);
 
-      expect(find.text('1翻'),
+      expect(find.text('2翻'),
           findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
+      expect(find.text('40符'), findsOneWidget);
 
-      expect(find.text('子：300点'), findsOneWidget);
-      expect(find.text('親：500点'), findsOneWidget);
+      expect(find.text('2600点'), findsOneWidget);
 
-      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('基本：30'), findsOneWidget);
       expect(find.text('面子：8'), findsOneWidget);
       expect(find.text('アタマ：0'), findsOneWidget);
       expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
+      expect(find.text('ツモ：0'), findsOneWidget);
     });
   });
 }
