+++ test/m/2018-2019/regular/10/m_20181011_test_template.dart
@@ -1,0 +1,50 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:mahjong_score/data/fu/fus.dart';
+import 'package:mahjong_score/data/yaku/yaku.dart';
+
+import '../../../test_widget_util.dart';
+
+// 東3局 1本場：流局
+// expect(find.text('断么九'), findsOneWidget);
+
+void main() {
+  group('2018/10/11 1試合目', () {
+    testWidgets('東1局 0本場', (WidgetTester tester) async {
+      await tester.pumpWidget(
+        createTestProviderScope(
+          yakus: [
+            YakuId.reach,
+            YakuId.tsumo,
+          ],
+          dora: 2,
+          honba: 0,
+          menzen: true,
+          tsumo: true,
+          oya: false,
+          mentsu1: FuMentsu.shuntsu,
+          mentsu2: FuMentsu.shuntsu,
+          mentsu3: FuMentsu.shuntsu,
+          mentsu4: FuMentsu.chuchanpaiAnko,
+          atama: FuAtama.otakaze,
+          machi: FuMachi.ryanmen,
+        ),
+      );
+
+      expect(find.text('立直'), findsOneWidget);
+      expect(find.text('門前自摸'), findsOneWidget);
+
+      expect(find.text('4翻'),
+          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
+      expect(find.text('30符'), findsOneWidget);
+
+      expect(find.text('子：2000点'), findsOneWidget);
+      expect(find.text('親：4000点'), findsOneWidget);
+
+      expect(find.text('基本：20'), findsOneWidget);
+      expect(find.text('面子：4'), findsOneWidget);
+      expect(find.text('アタマ：0'), findsOneWidget);
+      expect(find.text('待ち：0'), findsOneWidget);
+      expect(find.text('ツモ：2'), findsOneWidget);
+    });
+  });
+}
