--- test/m/2018-2019/regular/m_20181011_test.dart
@@ -1,878 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:mahjong_score/data/fu/fus.dart';
-import 'package:mahjong_score/data/yaku/yaku.dart';
-
-import '../../test_widget_util.dart';
-
-// 東3局 1本場：流局
-// expect(find.text('断么九'), findsOneWidget);
-
-void main() {
-  group('2018/10/11 1試合目', () {
-    testWidgets('東1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.pinfu,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('平和'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('3900点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東2局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.tsumo,
-            YakuId.pinfu,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: true,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('門前自摸'), findsOneWidget);
-      expect(find.text('平和'), findsOneWidget);
-
-      expect(find.text('4翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('20符'), findsOneWidget);
-
-      expect(find.text('子：1300点'), findsOneWidget);
-      expect(find.text('親：2600点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東3局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.sangenpai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-
-      expect(find.text('1翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('2000点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：2'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東3局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.tanyao,
-            YakuId.pinfu,
-            YakuId.ippatsu,
-          ],
-          dora: 1,
-          honba: 1,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('断么九'), findsOneWidget);
-      expect(find.text('平和'), findsOneWidget);
-      expect(find.text('一発'), findsOneWidget);
-
-      expect(find.text('5翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('12300点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東3局 2本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-          ],
-          dora: 1,
-          honba: 2,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.chuchanpaiAnko,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('4500点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東3局 3本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.tsumo,
-          ],
-          dora: 3,
-          honba: 3,
-          menzen: true,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.kanchan,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('門前自摸'), findsOneWidget);
-
-      expect(find.text('5翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('子：2300点'), findsOneWidget);
-      expect(find.text('親：4300点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    testWidgets('東4局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.kanchan,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('2600点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.sha,
-            YakuId.honitsu,
-            YakuId.toitoiho,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.chuchanpaiAnko,
-          mentsu2: FuMentsu.chuchanpaiAnko,
-          mentsu3: FuMentsu.chuchanpaiMinko,
-          mentsu4: FuMentsu.yaochuhaiMinko,
-          atama: FuAtama.sangenpai,
-          machi: FuMachi.shanpon,
-        ),
-      );
-
-      expect(find.text('西'), findsOneWidget);
-      expect(find.text('混一色'), findsOneWidget);
-      expect(find.text('対々和'), findsOneWidget);
-
-      expect(find.text('6翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('12000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：14'), findsOneWidget);
-      expect(find.text('アタマ：2'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南2局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.chitoitsu,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.none,
-          mentsu2: FuMentsu.none,
-          mentsu3: FuMentsu.none,
-          mentsu4: FuMentsu.none,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('七対子'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('25符'), findsOneWidget);
-
-      expect(find.text('3200点'), findsOneWidget);
-
-      expect(find.text('基本：25'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    // 南3局 0本場：流局
-
-    testWidgets('南3局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.nanBa,
-            YakuId.nanKaze,
-          ],
-          dora: 0,
-          honba: 1,
-          menzen: false,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.yaochuhaiMinko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('南（場）'), findsOneWidget);
-      expect(find.text('南（風）'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('2300点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：4'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南4局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.nanBa,
-            YakuId.nanKaze,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.chuchanpaiAnko,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.yaochuhaiAnko,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('南（場）'), findsOneWidget);
-      expect(find.text('南（風）'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('50符'), findsOneWidget);
-
-      expect(find.text('6400点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：12'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-  });
-
-  group('2018/10/11 2試合目', () {
-    testWidgets('東1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.ippatsu,
-            YakuId.tsumo,
-            YakuId.tanyao,
-          ],
-          dora: 2,
-          honba: 0,
-          menzen: true,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.tanki,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('一発'), findsOneWidget);
-      expect(find.text('門前自摸'), findsOneWidget);
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('6翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('子：3000点'), findsOneWidget);
-      expect(find.text('親：6000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    testWidgets('東2局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.chitoitsu,
-            YakuId.tsumo,
-          ],
-          dora: 2,
-          honba: 0,
-          menzen: true,
-          tsumo: true,
-          oya: true,
-          mentsu1: FuMentsu.none,
-          mentsu2: FuMentsu.none,
-          mentsu3: FuMentsu.none,
-          mentsu4: FuMentsu.none,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('門前自摸'), findsOneWidget);
-      expect(find.text('七対子'), findsOneWidget);
-
-      expect(find.text('5翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('25符'), findsOneWidget);
-
-      expect(find.text('4000点 all'), findsOneWidget);
-
-      expect(find.text('基本：25'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東2局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.pinfu,
-          ],
-          dora: 0,
-          honba: 1,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('平和'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('2300点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東3局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-          ],
-          dora: 2,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.yaochuhaiAnko,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-
-      expect(find.text('3翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('7700点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：8'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東3局 1本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-          ],
-          dora: 1,
-          honba: 1,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.chuchanpaiAnko,
-          mentsu2: FuMentsu.chuchanpaiAnko,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.baKaze,
-          machi: FuMachi.tanki,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('50符'), findsOneWidget);
-
-      expect(find.text('5100点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：8'), findsOneWidget);
-      expect(find.text('アタマ：2'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東3局 2本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.tanyao,
-            YakuId.pinfu,
-          ],
-          dora: 0,
-          honba: 2,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('平和'), findsOneWidget);
-      expect(find.text('断么九'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('30符'), findsOneWidget);
-
-      expect(find.text('2600点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('東4局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.haku,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.yaochuhaiAnko,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('白'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('2600点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：8'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南1局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.tanyao,
-            YakuId.pinfu,
-            YakuId.tsumo,
-            YakuId.ippatsu,
-            YakuId.sanshokudojun,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: true,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.shuntsu,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('門前自摸'), findsOneWidget);
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('平和'), findsOneWidget);
-      expect(find.text('断么九'), findsOneWidget);
-      expect(find.text('一発'), findsOneWidget);
-      expect(find.text('三色同順'), findsOneWidget);
-
-      expect(find.text('7翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('20符'), findsOneWidget);
-
-      expect(find.text('子：3000点'), findsOneWidget);
-      expect(find.text('親：6000点'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南2局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-          ],
-          dora: 1,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: false,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.yaochuhaiAnko,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-
-      expect(find.text('2翻'),
-          findsNWidgets(2)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('40符'), findsOneWidget);
-
-      expect(find.text('2600点'), findsOneWidget);
-
-      expect(find.text('基本：30'), findsOneWidget);
-      expect(find.text('面子：8'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    testWidgets('南3局 0本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.chitoitsu,
-            YakuId.honroto,
-          ],
-          dora: 0,
-          honba: 0,
-          menzen: true,
-          tsumo: false,
-          oya: true,
-          mentsu1: FuMentsu.none,
-          mentsu2: FuMentsu.none,
-          mentsu3: FuMentsu.none,
-          mentsu4: FuMentsu.none,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('七対子'), findsOneWidget);
-      expect(find.text('混老頭'), findsOneWidget);
-
-      expect(find.text('4翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('25符'), findsOneWidget);
-
-      expect(find.text('9600点'), findsOneWidget);
-
-      expect(find.text('基本：25'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-
-    // 南3局 1本場：流局
-
-    testWidgets('南4局 2本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.reach,
-            YakuId.tsumo,
-          ],
-          dora: 5,
-          honba: 2,
-          menzen: true,
-          tsumo: true,
-          oya: true,
-          mentsu1: FuMentsu.shuntsu,
-          mentsu2: FuMentsu.shuntsu,
-          mentsu3: FuMentsu.yaochuhaiAnkan,
-          mentsu4: FuMentsu.shuntsu,
-          atama: FuAtama.suhai,
-          machi: FuMachi.kanchan,
-        ),
-      );
-
-      expect(find.text('立直'), findsOneWidget);
-      expect(find.text('門前自摸'), findsOneWidget);
-
-      expect(find.text('7翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('60符'), findsOneWidget);
-
-      expect(find.text('6200点 all'), findsOneWidget);
-
-      expect(find.text('基本：20'), findsOneWidget);
-      expect(find.text('面子：32'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：2'), findsOneWidget);
-      expect(find.text('ツモ：2'), findsOneWidget);
-    });
-
-    testWidgets('南4局 3本場', (WidgetTester tester) async {
-      await tester.pumpWidget(
-        createTestProviderScope(
-          yakus: [
-            YakuId.chitoitsu,
-            YakuId.tsumo,
-          ],
-          dora: 2,
-          honba: 3,
-          menzen: true,
-          tsumo: true,
-          oya: false,
-          mentsu1: FuMentsu.none,
-          mentsu2: FuMentsu.none,
-          mentsu3: FuMentsu.none,
-          mentsu4: FuMentsu.none,
-          atama: FuAtama.suhai,
-          machi: FuMachi.ryanmen,
-        ),
-      );
-
-      expect(find.text('七対子'), findsOneWidget);
-      expect(find.text('門前自摸'), findsOneWidget);
-
-      expect(find.text('5翻'),
-          findsNWidgets(1)); // 1翻,2翻,3翻,6翻,役満は役選択部分にもテキストがあるため、2widgetヒットする
-      expect(find.text('25符'), findsOneWidget);
-
-      expect(find.text('子：2300点'), findsOneWidget);
-      expect(find.text('親：4300点'), findsOneWidget);
-
-      expect(find.text('基本：25'), findsOneWidget);
-      expect(find.text('面子：0'), findsOneWidget);
-      expect(find.text('アタマ：0'), findsOneWidget);
-      expect(find.text('待ち：0'), findsOneWidget);
-      expect(find.text('ツモ：0'), findsOneWidget);
-    });
-  });
-}
